% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datedName.R
\name{datedName}
\alias{datedName}
\title{Time stamped string}
\usage{
datedName(name, sep)
}
\arguments{
\item{name}{A name to concatenate as a suffix.}
}
\description{
Function for generating a string preceeded by a time stamp: `YYYYMMDD_HHMMSS_filename`.
Takes a character input
}
