% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datedSave.R
\name{datedSave}
\alias{datedSave}
\title{Save time stamped file}
\usage{
datedSave(input, save_function, name)
}
\arguments{
\item{input}{A data frame.}

\item{save_function}{A file format either "excel" or "robject"; must be class of character.}

\item{name}{A name for the file.}
}
\description{
A function for easily saving time stamped objects, and data frames.
Takes a variable r-object, one of the following save functions, "txt, "csv", "excel" or "robject", data frame or r-objects and then followed by a character name.
}
