% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_move_col.R
\name{df_move_col}
\alias{df_move_col}
\title{Move data frame columns with natural language}
\usage{
df_move_col(vec, movecommand)
}
\arguments{
\item{vec}{Get the names(df) of data frame; vector.}

\item{movecommand}{a string with the name of the column followed by the move command.}
}
\description{
A function for easily moving cols in df with natural language.
Takes the names() of the data frame and a move command in natural language; 4 options.
}
\details{
"after", "before", "first", "last"; are possible as string commands for movement. The column names should be input as a vector; use names(df) command.
}
\examples{
df_move_col(names(df), "colname20 first")
}
