% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_round_values.R
\name{df_round_values}
\alias{df_round_values}
\title{Function for rounding all values in a data frame keeping them in the numeric type.}
\usage{
df_round_values(df, digits)
}
\arguments{
\item{df}{a data frame input.}

\item{digits}{a number to which the values will be rounded to.}
}
\description{
Takes two variables, a data frame and a number to round to.
}
