% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_save_dated.R
\name{df_save_dated}
\alias{df_save_dated}
\title{A function for easily saving time stamped data frames.}
\usage{
df_save_dated(df, save_function, name)
}
\arguments{
\item{df}{A data frame}

\item{save_function}{A file format either "excel" or "robject"; must be class of character}

\item{name}{A name for the file}
}
\description{
Takes one of two save functions, "excel" or "robject", data frame, and name
}
