% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfDeconcat.R
\name{dfDeconcat}
\alias{dfDeconcat}
\title{Deconcatenate rows of a data frame}
\usage{
dfDeconcat(df, pattern)
}
\arguments{
\item{df}{A data frame type object.}

\item{pattern}{A string to use a separator for the rows.}
}
\description{
Made for deconcatenating the rows of the DESeq2 results object. Works well on other data frames, feed the data frame and the pattern to match.
}
\details{
It duplicates the data for each row and sets the separated names values to rownames, separately.
}
