% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfRmvNA.R
\name{dfRmvNA}
\alias{dfRmvNA}
\title{Remove NA rows from data frame.}
\usage{
dfRmvNA(df, col = "")
}
\arguments{
\item{df}{A data frame type object.}

\item{col}{A string for selecting a column.}
}
\description{
Reliably removes all rows that include NA values for a given column in a data frame object.
}
\details{
Was made to work with res object of DESeq2.
}
