% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfRmvStr.R
\name{dfRmvStr}
\alias{dfRmvStr}
\title{Regex replace/remove from all rows of data frame}
\usage{
dfRmvStr(df, str1 = "", str2 = "", str3 = "", rep1 = "", rep2 = "", rep3 = "")
}
\arguments{
\item{df}{A data frame type input.}

\item{str1}{Regex to search for.å}

\item{str2}{Regex to search for.}

\item{str3}{Regex to search for.}

\item{rep1}{Replacement string.}

\item{rep2}{Replacement string.}

\item{rep3}{Replacement string.å}
}
\description{
A function for easily replacing/removing up to 3 different regular expressions from all rows in a data frame.
Takes a data frame object and outputs a data frame object. All regex set to `strn = ""` by default.
}
