% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfRound.R
\name{dfRound}
\alias{dfRound}
\title{Round all values in a data frame}
\usage{
dfRound(df, digits)
}
\arguments{
\item{df}{a data frame input.}

\item{digits}{a number to which the values will be rounded to.}
}
\description{
Function for rounding all values in a data frame keeping them in the numeric type.
Takes two variables, a data frame and a number to round to.
}
