% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDots.R
\name{dfUnevenCheck}
\alias{dfUnevenCheck}
\title{Compare uneven data frames}
\usage{
dfUnevenCheck(net1, net2)
}
\arguments{
\item{net1}{A data frame type object, of any height; must be equal in width to the second input.}

\item{net2}{A data frame type object, of any size; must be equal in width to the first input.}
}
\description{
Compare two data frames of differing heights.
}
\details{
Good for scoring two data frame of different heights; normally on obtains a truth table and can score the similarity between two data frames by using `df1 == df2`. When they are not of equal heights this is not possible.

Here I used the `%in%` operator to check the similarity between two uneven data frames and return a pseudo score along with a pseudo truth table.

Note that scores and truth tables returned are not guaranteed to be representative of the similarity of the two data frames input. This is because of the nature of the `%in%` operator.

A sanitation check is done to prevent the input of two objects of different widths.
}
