% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotMerge.R
\name{dotMerge}
\alias{dotMerge}
\title{Merge CARNIVAL dot networks}
\usage{
dotMerge(dot1, dot2)
}
\arguments{
\item{dot1}{A .dot network network.}
}
\description{
Merges two CARNIVAL networks, removing any EXACT interactions that are in double.
}
\details{
Use readLines() to read the networks as a string object, each line should be a new element in an array.
}
