% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotMinify.R
\name{dotMinify}
\alias{dotMinify}
\title{Minify CARNIVAL network to list matches}
\usage{
dotMinify(file = "", matchList = "")
}
\arguments{
\item{file}{A .dot network graph file from CARNIVAL.}

\item{matchList}{A vector list of gene names as strings to match for; reduce the network to sub networks.}
}
\description{
Takes a .dot network file and vector list of gene names, and matches the node pairs (origin target) to the list.
}
\details{
Adds upstream interactions and downstream interactions to the matching list to get the largest subnetworks and all indirect interactions to the gene of interest.
}
