% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveCellVenn.R
\name{fiveCellVenn}
\alias{fiveCellVenn}
\title{Create five cell venn diagram}
\usage{
fiveCellVenn(
  colorList = col2rgb(rainbow(5)),
  cellCounts = seq(1, 31, 1),
  cellLabels = c("one", "two", "three", "four", "five"),
  saturation = 0.25
)
}
\description{
Prototype taken from: https://stackoverflow.com/questions/32440128/nice-looking-five-sets-venn-diagrams
}
