% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileWd.R
\name{getFileWd}
\alias{getFileWd}
\title{Get path of current file}
\usage{
getFileWd()
}
\description{
A function for easily getting the path of the current file you're working in.
}
\details{
This function can only be executed from the editor, not the console; it uses rstudioapi. It will not work independently of R-studio.
Takes no inputs, and can have it's string results set to a variable.
}
