% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDefined.R
\name{isDefined}
\alias{isDefined}
\title{Check if variable is defined}
\source{
https://stackoverflow.com/questions/9368900/how-to-check-if-object-variable-is-defined-in-r
}
\usage{
isDefined(var)
}
\arguments{
\item{var}{variable to check; no quotations.}
}
\description{
Easily check if a variable has been previously instantiated and defined. Does not require quotations.
}
\examples{
isDefined(a) # FALSE
a <- 10
isDefined(a) # TRUE
}
