% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Load .RData into var}
\usage{
loadRData(file)
}
\description{
Allows to load an .RData file into a variable name, not the global environment.
}
\details{
Experimental; unstable.
}
