% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDot.R
\name{parseDot}
\alias{parseDot}
\title{Parse .dot graph files to data frames}
\usage{
parseDot(file)
}
\arguments{
\item{file}{A file path as a string.}
}
\description{
A function for parsing .dot type files (specifically those produced by the saez/CARNIVAL package), into data frames.
}
\details{
Good for analysing and comparing results; in human readable formats or otherwise.

Takes a file path or name as a string; be sure to be in the correct working directory.
Sets and instatiates a class, returns to variable; use `@` to access the sets relating to this functions results.
}
