% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reset_par.R
\name{plot_reset_par}
\alias{plot_reset_par}
\title{Reset margins on plots}
\usage{
plot_reset_par()
}
\description{
A function for reseting margins for plots.
}
\details{
Has no inputs just run the function.
}
\examples{
reset_par()
}
