% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetPar.R
\name{resetPar}
\alias{resetPar}
\title{Reset margins on plots}
\usage{
resetPar()
}
\description{
A function for reseting margins for plots.
}
\details{
Has no inputs just run the function.
}
\examples{
resetPar()
}
