% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{Function for rounding all values in a data frame keeping them in the numeric type.}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{a data frame input.}

\item{digits}{a number to which the values will be rounded to.}
}
\description{
Takes two variables, a data frame and a number to round to.
}
