% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dated_df.R
\name{save_dated_df}
\alias{save_dated_df}
\title{A function for easily saving time stamped data frames.}
\usage{
save_dated_df(save_function, df, name)
}
\arguments{
\item{save_function}{A file format either "excel" or "robject"; must be class of character}

\item{df}{A data frame}

\item{name}{A name for the file}
}
\description{
Takes one of two save functions, "excel" or "robject", data frame, and name
}
