% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servrKnitRhtml.R
\name{servrKnitRhtml}
\alias{servrKnitRhtml}
\title{Knit .Rhtml launch localhost; show in viewer}
\usage{
servrKnitRhtml(
  save = TRUE,
  input = "index.Rhtml",
  clear_cache = TRUE,
  wd = "",
  port = "6969"
)
}
\arguments{
\item{save}{set to save the current working file by default; this easily allows for you to set this function to a hotkey.}

\item{input}{set to "index.Rhtml" by default.}

\item{clear_cache}{set to TRUE by default.}

\item{wd}{set to "" by default, if left will recover the wd of the current working file.}

\item{port}{set to "6969" by default, allows you to set the port on localhost.}

\item{output}{set to "index.html" by default.}
}
\description{
This function Knits an .Rhtml file to .html, within the current working directory of the working file, if no wd specified. It gets this working directory by the `getFileWd()` function of `octavius`.
}
\details{
A localhost server is launched on `port: 6969`; port forward to make this server externally accessible: `http://localhost:6969`

A nice feature is to set this function to `cmd + shift + s` as this could have it knit and preview automatically on save.

Finally the local viewer (R-studio's) is launched; note that the default preview window for knitting .Rhtml is not launched.
}
