% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBinTable.R
\name{toBinTable}
\alias{toBinTable}
\title{Convert TRUE/FALSE table to binary table}
\usage{
toBinTable(df, convertNA = 1)
}
\arguments{
\item{df}{A data frame object of TRUE/FALSE values; may also contain NAs.}

\item{convertNA}{A numeric value to which convert NAs.}
}
\description{
Converts a data frame object that contains TRUE and FALSE; also NAs to a binary table.
}
\details{
Good for subsequently passing onto the `venn` package for visualisation.

A sanitation check is done on the convertNA input to ensure it the resulting table is indeed binary.
}
