% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topSigExtract.R
\name{topSigExtract}
\alias{topSigExtract}
\title{Extract top statistically significant lg2fc genes}
\usage{
topSigExtract(res, alpha = 0.1)
}
\arguments{
\item{res}{A DESeq2 results type object; data frame type object.}

\item{alpha}{Alpha threshold, default set to 0.1.}
}
\description{
Extracts all top significantly differentially expressed genes from res object of DESeq2. The results will be automatically ordered per lowest alpha, then reordered per highest absolute value for the lg2fc column  of the res object.
}
\details{
Allows for setting an alpha threshold. Note that this function will use the dfRmvNA function to remove any NA rows found in the results object. Note that NA results in the padj column of the DESeq2 results object are normal; consult the DESeq2 documentation.

Was made to work with res object of DESeq2.
}
