% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_unique_check.R
\name{vec_unique_check}
\alias{vec_unique_check}
\title{Vector values unique check}
\usage{
vec_unique_check(vector)
}
\arguments{
\item{vec}{A vector of any kind, can contain logicals or characters etc.}
}
\description{
A function that will take a vector and test if the values are unique.
}
