% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Volcano plot}
\usage{
volcanoplot(
  df,
  pcutoff = 0.05,
  lcutoff = 1,
  nlabs = 10,
  xlim = NULL,
  ylim = NULL,
  ptype = 20,
  psize = 1,
  main = "Volcano plot\\n Cutoffs: padj < 0.05, abs(lg2fc) > 1",
  sub = "",
  ldispmethod = "SANN",
  lcex = 1,
  lpos = 2,
  loffset = 0.75,
  xnsig = 0,
  ynsig = -10,
  nsigcex = 1
)
}
\description{
Creates simple volcano plot based on base R graphics. Highly customisable and allows for deeper arguments.
}
\details{
Takes a minimal data frame; col-1 log2FoldChange, col-2 padj. Padj can contain NAs these will be removed by internal function.

Note that the function ALSO returns an object with the differentially expressed genes, rownames are genes, column one is log2FoldChange and column two is the padj values. Simply set the function to a variable to save the object.
}
