% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dated_name.R
\name{dated_name}
\alias{dated_name}
\title{Time stamped string}
\usage{
dated_name(name)
}
\arguments{
\item{name}{A name to concatenate as a suffix.}
}
\description{
Function for generating a string preceeded by a time stamp: `YYYYMMDD_HHMMSS_filename`.
Takes a character input
}
