% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_move_col.R
\name{df_move_col}
\alias{df_move_col}
\title{A function for easily moving cols in df with natural language}
\usage{
df_move_col(invec, movecommand)
}
\arguments{
\item{invec}{the names(df) for a data frame}

\item{movecommand}{a string with the name of the column followed by the move command}
}
\description{
Takes the names() of the data frame and a move command in natural language; 4 options
}
\details{
"after", "before", "first", "last"; are possible as string commands for movement.
}
\examples{
df_move_col(names(df), "colname20 first")
df[df_move_col(names(df), "colname20 first")]
}
