% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Volcano plot}
\usage{
volcanoplot(
  df,
  pcutoff = 0.05,
  lcutoff = 1,
  nlabs = 10,
  xlim = NULL,
  ylim = NULL,
  ptype = 20,
  psize = 1,
  main = "Volcano plot\\n Cutoffs: padj < 0.05, abs(lg2fc) > 1",
  sub = "",
  ldispmethod = "SANN",
  lcex = 1,
  lpos = 2,
  loffset = 0.75,
  xnsig = 0,
  ynsig = -10,
  nsigcex = 1
)
}
\arguments{
\item{df}{A data frame type object first column log2FoldChange, column two is padj.}

\item{pcutoff}{Default set to 0.05, these are p adjusted values, function originally made to work with DESeq2 results which uses an FDR correction. Check you values compatibility for correct usage.}

\item{lcutoff}{Default set to 1, this is the absolute value at which the log2FoldChange will be cutoff. 1 Is the defaulte but when working with log2FoldChange 0.5 can be biologically relevant as it is indeed a 41\% increase in expression over the controls.}

\item{nlabs}{Default set to 10, this is the number of points that will have labels associated to them on the plot. Reduce the number to avoid overlapping labels and only set those that are most over/under-expressed AND are statistically significant. Internal function topSigExtract() used.}

\item{xlim}{Default set to NULL, this is the x axis limits of the plot. Plot first then set according to needs.}

\item{ylim}{Default set to NULL, this is the y axis limits of the plot. Plot first then set according to needs.}

\item{ptype}{Default set to 20, this is the type of points used on the plot. Default is round circle.}

\item{psize}{Default set to 1cm. This is the point size on the plot.}

\item{main}{Default set to "Volcano plot Cutoffs: padj < 0.05, abs(lg2fc) > 1".}

\item{sub}{Default set to empty string "".}

\item{ldispmethod}{Default set to "SANN", using the "maptools" library, "SANN" or "GA" are possible. This is the algorithm by which the point labels are dispersed to avoid overlaps.}

\item{lcex}{Default set to 1, this is the point labels size.}

\item{lpos}{Default set to 2, this is the point label position relative to the point.}

\item{loffset}{Default set to 0.75, this is the point label offset from the point; or distance.}

\item{xnsig}{Default set to 0, this is the x position for the plot information text; it states the number of DEGs and the parameters used to determine cutoffs.}

\item{ynsig}{Default set to -10, this is the y position for the plot information text; it states the number of DEGs and the parameters used to determine cutoffs.}

\item{nsigcex}{Default set to 0, this is the text size for the plot information text; it states the number of DEGs and the parameters used to determine cutoffs.}
}
\description{
#' Creates simple volcano plot based on base R graphics. Highly customisable and allows for deeper arguments.
}
\details{
Takes a minimal data frame; col-1 log2FoldChange, col-2 padj. Padj can contain NAs these will be removed by internal function.

Note that the function ALSO returns an object with the differentially expressed genes, rownames are genes, column one is log2FoldChange and column two is the padj values. Simply set the function to a variable to save the object.
}
